[FEATURE] Make standalone and add optional backends
This change makes the PhpProfiler independent of Plumber, so it can be
used with having to install Plumber as well.

This is useful if you want to separate the UI for viewing profiling
results from the actual data collection.

In addition this change makes it possible to store collected profiling
data into XHGui and XHProf.io.

Basic documentation has been added.